fix 접근성 수정
