+++ src/front/package.json
@@ -20,6 +20,7 @@
     "react-dom": "^18.2.0",
     "react-router-dom": "^6.9.0",
     "react-uuid": "^2.0.0",
+    "rollup-plugin-visualizer": "^5.9.0",
     "styled-components": "^5.3.9",
     "styled-reset": "^4.4.5"
   },
@@ -32,6 +33,7 @@
     "cypress": "^12.10.0",
     "typescript": "^4.9.3",
     "vite": "^4.2.0",
-    "vite-plugin-favicon": "^1.0.8"
+    "vite-plugin-favicon": "^1.0.8",
+    "webpack-bundle-analyzer": "^4.8.0"
   }
 }
