+++ src/front/public/index.html
@@ -4,6 +4,27 @@
     <meta charset="UTF-8" />
     <link rel="icon" type="image/svg+xml" href="/vite.svg" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <meta
+      name="description"
+      content="if you want to play tennis, you can find a partner here."
+    />
+    <meta name="keywords" content="tennis, partner, tennis partner" />
+    <meta name="author" content="TennisPartner" />
+    <meta name="theme-color" content="#ffffff" />
+    <meta name="twitter:card" content="summary_large_image" />
+    <meta name="twitter:site" content="@tennis_partner" />
+    <meta name="twitter:creator" content="@tennis_partner" />
+    <meta name="twitter:title" content="테니스 파트너" />
+    <meta name="twitter:description" content="같이 테니스 치실래요?" />
+    <meta name="twitter:image" content="%PUBLIC_URL%/og-image.png" />
+    <meta name="twitter:image:alt" content="테니스 파트너" />
+    <meta
+      name="twitter:domain"
+      content="https://front-deploy-three.vercel.app/"
+    />
+    <meta name="twitter:app:name:iphone" content="테니스 파트너" />
+    <meta name="twitter:app:name:ipad" content="테니스 파트너" />
+    <meta name="twitter:app:name:googleplay" content="테니스 파트너" />
 
     <meta property="og:type" content="website" />
     <meta property="og:url" content="https://front-deploy-three.vercel.app/" />
