+++ src/front/src/components/Auth/AuthButton.tsx
@@ -3,10 +3,15 @@ import styled from "styled-components";
 interface AuthButtonProps {
   children: string;
   onClick?: (e: any) => void;
+  style?: any;
 }
 
-const AuthButton = ({ children, onClick }: AuthButtonProps) => {
-  return <CustomButton onClick={onClick}>{children}</CustomButton>;
+const AuthButton = ({ children, onClick, style }: AuthButtonProps) => {
+  return (
+    <CustomButton style={style} onClick={onClick}>
+      {children}
+    </CustomButton>
+  );
 };
 
 const CustomButton = styled.button`
