+++ src/front/src/components/Footer.tsx
@@ -1,15 +1,11 @@
 import React, { useContext, useEffect, useState } from "react";
 import styled from "styled-components";
-import { useLocation } from "react-router-dom";
-import { Link } from "react-router-dom";
+import { useLocation, Link } from "react-router-dom";
 import { checkLoginState } from "../util/checkLoginState";
 import { userContext } from "../context/userContext";
-import axios from "axios";
 
 const Footer = () => {
   const location = useLocation();
-  const baseUrl = import.meta.env.VITE_APP_BACK_END_AWS;
-  const accessToken = localStorage.getItem("accessToken");
 
   const [loginState, setLoginState] = useState(false);
