+++ src/front/src/pages/myPage/index.tsx
@@ -5,7 +5,9 @@ import AuthButton from "../../components/Auth/AuthButton";
 
 import { useNavigate } from "react-router-dom";
 import axios from "axios";
-import Compressor from "compressorjs";
+
+import { compressImage } from "../../util/compressImage";
+import { dataURLtoFile } from "../../util/dataURLtoFile";
 
 import { userContext } from "../../context/userContext";
 
@@ -13,12 +15,17 @@ const MyPage = () => {
   const [nickName, setNickName] = useState("");
   const [gender, setGender] = useState("");
   const [ntrp, setNtrp] = useState("");
-  const [isFinished, setIsFinished] = useState(false);
+
+  const [userProfile, setUserProfile] = useState({
+    nickName: "",
+    gender: "",
+    ntrp: "",
+  });
 
   const { user, setUser }: any = useContext(userContext);
   const navigate = useNavigate();
 
-  const defaultProfile = "profile.png";
+  const defaultProfile = "profile.webp";
 
   const baseUrl = import.meta.env.VITE_APP_BACK_END_AWS;
 
@@ -44,36 +51,6 @@ const MyPage = () => {
       });
   };
 
-  const compressImage = (file: any, callback: any) => {
-    new Compressor(file, {
-      quality: 0.6, // 이미지 품질
-      maxWidth: 300, // 이미지 최대 너비
-      maxHeight: 300, // 이미지 최대 높이
-      success(result) {
-        const reader = new FileReader();
-        reader.readAsDataURL(result);
-        reader.onloadend = () => {
-          callback(reader.result);
-        };
-      },
-      error(err) {
-        console.error(err.message);
-      },
-    });
-  };
-
-  const dataURLtoFile = (dataurl: any, filename: any) => {
-    const arr = dataurl.split(",");
-    const mime = arr[0].match(/:(.*?);/)![1];
-    const bstr = atob(arr[1]);
-    let n = bstr.length;
-    const u8arr = new Uint8Array(n);
-    while (n--) {
-      u8arr[n] = bstr.charCodeAt(n);
-    }
-    return new File([u8arr], filename, { type: mime });
-  };
-
   const takePicture = () => {
     if (videoRef.current && canvasRef.current) {
       const canvas = canvasRef.current;
@@ -96,6 +73,7 @@ const MyPage = () => {
     }
   };
 
+  // 로그아웃
   const logout = () => {
     localStorage.removeItem("accessToken");
     localStorage.removeItem("refreshToken");
@@ -115,15 +93,12 @@ const MyPage = () => {
         },
         {
           headers: {
+            contentType: "application/json",
             Authorization: `Bearer ${accessToken}`,
           },
         }
       )
       .then((res) => {
-        setIsFinished(true);
-        setTimeout(() => {
-          setIsFinished(false);
-        }, 3000);
         return res;
       })
       .catch((err) => {
@@ -142,6 +117,7 @@ const MyPage = () => {
       const result = await axios
         .get(`${baseUrl}/login/api/users`, {
           headers: {
+            contentType: "application/json",
             Authorization: `Bearer ${accessToken}`,
           },
         })
@@ -168,7 +144,11 @@ const MyPage = () => {
 
   return (
     <CreateProfileContainer>
-      <LogoutButton onClick={logout}>로그아웃</LogoutButton>
+      {!isVideo ? (
+        <LogoutButton onClick={logout}>로그아웃</LogoutButton>
+      ) : (
+        <LogoutButton onClick={takePicture}>사진 찍기</LogoutButton>
+      )}
       <h1>내 정보 등록</h1>
       <ProfilePicture onClick={startVideo}>
         {isVideo ? (
@@ -180,20 +160,7 @@ const MyPage = () => {
           <img src={picture ? picture : defaultProfile} alt="프로필 사진" />
         )}
       </ProfilePicture>
-      {isVideo && (
-        <button
-          style={{
-            width: "100px",
-            height: "32px",
-            borderRadius: "10px",
-            backgroundColor: "#FFC0CB",
-            border: "1px solid #FFC0CB",
-          }}
-          onClick={takePicture}
-        >
-          사진 찍기
-        </button>
-      )}
+
       <NickNameBox>
         <AuthInput
           value={nickName}
@@ -226,12 +193,12 @@ const MyPage = () => {
         </GenderCheck>
       </GenderBox>
       <NTRPBox>
-        <h2>NTRP</h2>
+        <label htmlFor="ntrp">NTRP</label>
         <NTRPCheck>
           <input
             type="range"
-            id="volume"
-            name="volume"
+            id="ntrp"
+            name="ntrp"
             min="0"
             max="5"
             step="0.5"
@@ -241,8 +208,9 @@ const MyPage = () => {
           <span>{ntrp}</span>
         </NTRPCheck>
       </NTRPBox>
-      <AuthButton onClick={postUserInfo}> 수정하기 </AuthButton>
-      {/* {isFinished && <FinishMark> 변경 완료 </FinishMark>} */}
+      <AuthButton style={{ marginTop: "0px" }} onClick={postUserInfo}>
+        수정하기
+      </AuthButton>
     </CreateProfileContainer>
   );
 };
@@ -270,11 +238,10 @@ const CreateProfileContainer = styled.div`
   flex-direction: column;
   align-items: center;
   justify-content: center;
-  gap: 32px;
+  gap: 16px;
 
-  height: calc(100vh - 80px);
-
-  overflow: auto;
+  width: 100%;
+  height: 100vh;
 
   h1 {
     display: flex;
@@ -286,21 +253,6 @@ const CreateProfileContainer = styled.div`
     line-height: 39px;
     justify-content: center;
   }
-`;
-
-const FinishMark = styled.div`
-  display: flex;
-  align-items: center;
-  justify-content: center;
-
-  width: 300px;
-
-  font-style: normal;
-  font-weight: 500;
-  font-size: 16px;
-  line-height: 20px;
-
-  color: ${({ theme }) => theme.colors.messageError};
 `;
 
 const NickNameBox = styled.div`
@@ -347,13 +299,10 @@ const GenderCheck = styled.div`
 const NTRPBox = styled.div`
   display: flex;
   align-items: left;
-  justify-content: center;
-  gap: 10px;
 
   width: 300px;
-  margin-bottom: 20px;
 
-  h2 {
+  label {
     color: #6b6b6b;
 
     font-size: 32px;
