+++ src/front/src/util/dataURLtoFile.ts
@@ -1,0 +1,12 @@
+// dataURL을 File로 변환
+export const dataURLtoFile = (dataurl: any, filename: any) => {
+  const arr = dataurl.split(",");
+  const mime = arr[0].match(/:(.*?);/)![1];
+  const bstr = atob(arr[1]);
+  let n = bstr.length;
+  const u8arr = new Uint8Array(n);
+  while (n--) {
+    u8arr[n] = bstr.charCodeAt(n);
+  }
+  return new File([u8arr], filename, { type: mime });
+};
