+++ src/front/vite.config.ts
@@ -1,11 +1,17 @@
 import { defineConfig } from "vite";
 import react from "@vitejs/plugin-react";
+import { visualizer } from "rollup-plugin-visualizer";
 
 // https://vitejs.dev/config/
 export default defineConfig({
   plugins: [
     react({
       include: "**/*.tsx",
+    }),
+    visualizer({
+      filename: "./dist/report.html",
+      open: true,
+      brotliSize: true,
     }),
   ],
 });
